Remove non-default VideoEncoder::EncoderInfo() ctor
This removes the constructor that takes all members as parameters.
A main goal with this struct is to make it easy to add or remove fields
without having to refactor implementations and usage all over the palce,
this breaks down if we need to update the arguments in this constructor
every time.

Also cleans up a few commented-out lines that were forgotten in an
earliers cl.

Bug: webrtc:9722
Change-Id: I22db35d399469bdb3bfd8bdc531b24ead9fe1c3e
Reviewed-on: https://webrtc-review.googlesource.com/c/109570
Reviewed-by: Niels Moller <nisse@webrtc.org>
Reviewed-by: Per Kjellander <perkj@webrtc.org>
Commit-Queue: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#25542}