Don't include packetization overhead in protection bitrate.
If we do, the bitrate allocator will assume there can be a lot a FEC
and other things and bumps the max probing bitrate by 2x.
This caused a bunch of perf tests to change in a non-obvious way.

This is a follow-up to
https://webrtc-review.googlesource.com/c/src/+/115410

Bug: webrtc:10155, chromium:922396
Change-Id: I51d3611cb21d98a8fab1bfab2d8f167ed859696d
Reviewed-on: https://webrtc-review.googlesource.com/c/118043
Reviewed-by: Niels Moller <nisse@webrtc.org>
Commit-Queue: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#26319}