Deprecate FirstSubFrameInFrame() and LastSubFrameInFrame()
In preparation for adding a discardability flag in
RtpGenericFrameDescriptor, deprecate two bits which are always
in practice set to TRUE.

This is conceptual deprecation. RTC_DEPRECATED cannot actually be
applied, because we still want to be able to parse those bits
and make sure they are truly set to TRUE when TRUE is expected.

Bug: webrtc:10214
Change-Id: I7d6cb640fe27f142578883389cc67d326c90f7bb
Reviewed-on: https://webrtc-review.googlesource.com/c/118381
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Commit-Queue: Elad Alon <eladalon@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#26340}