Create a composite implementation of RtpTransportInternal.
This will be used to multiplex multiple transports during SDP
negotiation.  When the offerer watns to support multiple RTP transports,
it will combine them into a singla CompositeRtpTransport.

CompositeRtpTransport can receive from any of the offered transports
while waiting for an answer to arrive.

The choice of which transport is used to send must be driven by the SDP
answer.  If a provisional answer arrives, the composite can be set to
send using the chosen transport, while maintaining other transports in
case the peer changes its mind.  When the final answer arrives, the
composite will be deleted and replaced with the chosen transport.

Bug: webrtc:9719
Change-Id: Ib8cea77ef202f37086723bfa2c71e2aa5995a912
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/138281
Commit-Queue: Bjorn Mellem <mellem@webrtc.org>
Reviewed-by: Anton Sukhanov <sukhanov@webrtc.org>
Reviewed-by: Steve Anton <steveanton@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#28093}