Specify min_sdk_version for unittest apks also in GN configs
This is a followup to https://webrtc-review.googlesource.com/142235

This was missed and it fails the new assert in https://chromium-review.googlesource.com/1675323
Example of failing roll: https://webrtc-review.googlesource.com/143844

This comes with the same caveats, i.e. the information is duplicated in AndroidManifest. But the new assert prevents de-sync between them :)

The version is being specified unconditionally because so is AndroidManifest.

Bug: chromium:891996
Change-Id: I92a2011c8fda55849bd9ba8df619f897f48e58da
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/143795
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Commit-Queue: Oleh Prypin <oprypin@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#28378}