Only combine media transport and ICE states if used for media.
Media transport (or, equivalently, datagram transport) may only be
created for data channels.  In this case, it's not appropriate to
consider ICE not-yet-connected or failed due to the media transport's
state.  If the media transport disconnects or fails, it will signal data
channels separately.

Bug: webrtc:9719
Change-Id: Ieb7cb307116e479d01616559d8bafdfc650a78c5
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/149420
Reviewed-by: Benjamin Wright <benwright@webrtc.org>
Reviewed-by: Steve Anton <steveanton@webrtc.org>
Commit-Queue: Bjorn Mellem <mellem@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#28884}