Add multi head queue implementation.
Queue with multiple heads is planned to be used in
DefaultVideoQualityAnalyzer to store stream state. Stream state contains
ordered sequence of frame ids that were send for this video stream.
When frame is received by one receiver it should be removed from state
for that receiver and kept for others.

How it is used can be found in this CL:
https://webrtc-review.googlesource.com/c/src/+/176411

Bug: webrtc:11631
Change-Id: Ic7fabf4d77131805a91f08a2ccfffc73c08d3e2d
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/176402
Commit-Queue: Artem Titov <titovartem@webrtc.org>
Reviewed-by: Karl Wiberg <kwiberg@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#31444}