Add QueryCodecSupport() to VideoEncoderFactory and VideoDecoderFactory
QueryCodecSupport() is needed in the coming work with extending
the MediaCapabilities API to WebRTC. In the API, a user can query
whether a specified media configuration is supported, power
efficient, and smooth. QueryCodecSupport() will be used to determine
if the configuration is supported and power efficient.

Bug: chromium:1187565
Change-Id: Ib1d93433a180b433f0bf60d6e871d03cebc4c0a6
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/215922
Commit-Queue: Johannes Kron <kron@webrtc.org>
Reviewed-by: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#33843}