Revert "Propagate socket write errors for DtlsTransport"
There is a suspicion that it causes OpenSSL errors:
  [openssl_stream_adapter.cc(961)]
       OpenSSLStreamAdapter::Error(SSL_write, 5, 0)

This commit does change the interaction with OpenSSL as propagating the
socket write errors as SR_BLOCK results in calling BIO_set_retry_write,
as part of current implementation of OpenSSLStreamAdapter.

Testing this regression has proven to be hard to do manually.

This reverts commit edfaaef086ccff2dbff29d64c9a8d9f633637c57.

Bug: webrtc:12943
Change-Id: Ib6767bd4af68c59fd3b7cb051341876f175bb921
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/230420
Reviewed-by: Tommi <tommi@webrtc.org>
Commit-Queue: Victor Boivie <boivie@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#34875}