p2p: Add field trial WebRTC-SetSocketReceiveBuffer
This field trial sets a non-zero receive buffer on the media UDP socket
with the intention to result in less packet loss in situations when the
application can't read packets fast enough from the socket. This can be
due to e.g. external factors, e.g. operating system not scheduling the
application for a longer time, or due to internal factors, e.g. slow
processing, a long running garbage collector, and more.

The size as set as the field trial parameter, as e.g.
WebRTC-SetSocketReceiveBuffer/Enabled-250/ to set it to 250kb.

Bug: webrtc:13753
Change-Id: Iae38d0db0c595d6e0148a2fdeb85ee8895e90560
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/252581
Reviewed-by: Jonas Oreland <jonaso@webrtc.org>
Commit-Queue: Victor Boivie <boivie@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#36080}