Fix generate_licenses Presubmit when working within chromium
When working within chromium, we need to move up two more levels before
we can import the "find_depot_tools" module. Previously this was
managed by a loop that kept moving up until it found a directory named
'src', but this didn't work if the base-level webrtc code was not in
a directory named 'src'.

Bug: webrtc:14411
Change-Id: I2729603d6d2fd8242a1294ae5d95d946db619d0a
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/273561
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#37971}